refactor(sofi): extract SoFiTradeRealHwTest helpers into SoFiTestHelpers.kt
Pure mechanical extraction in preparation for the Phase 8 cross-device
tests.  No behaviour change: SoFiTradeRealHwTest's 9-step flow is
preserved exactly; the helpers it relies on now live in a top-level
internal helper file so SoFiCrossDeviceOwnerTest /
SoFiCrossDeviceTraderTest (Phase 8) can share the same bridge
boilerplate, bootstrap-readiness pollers, and per-route framing
without duplicating any of it.

Split:
- Top-level file-level helpers (stateless): packArgs,
  decodeIngressEnvelope, routerInvoke, routerQuery, appStateValue,
  waitForAppRouter, pollBalanceUntilTrustReady, getBalance,
  claimFaucetEra (NEW — port of the AndroidLayerProofTest:880-892
  pattern via the high-level NativeBoundaryBridge), u128be,
  u128beToLong, b32, blake3Like32, freshOutputTokenForRun.
- SoFiTestContext class (token-pair-state-bound): createAmmVault,
  publishRoutingAdvertisement, syncVaultsForPair, findAndBindBestPath,
  signRouteCommit, computeExternalCommitment, publishExternalCommitment,
  unlockVaultRouted, listOwnedAmmVaults.

Constants (TAG, INPUT_TOKEN, INITIAL_RESERVE_A/_B, INPUT_AMOUNT,
MIN_ERA_BALANCE, FAUCET_CLAIM_AMOUNT, VAULT1/2_FEE_BPS, MAX_PATHS,
SLIPPAGE_BPS, FLOOR_BPS, RESERVE_POLL_ATTEMPTS, SPIN_BUDGET_PER_POLL)
moved to top-level so all three tests reference the same values.

SoFiTradeRealHwTest itself drops from 698 lines to ~245, becoming a
thin driver that orchestrates the 9-step flow.  Any future
bridge-pattern fix lands in one file instead of three.

Build verified: ./gradlew :app:assembleDebugAndroidTest passes clean.
Phase 5 single-device behaviour preserved (hardware re-run pending).
